Added package VLF (very large file).
This package is used to open large files instantly; it works by opening
files in batches. This allows opening files that may be terabytes long.

In Exordium a large file is a file that is above 100MB (see
init-look-and-feel.el). Emacs asks confirmation before opening such as
file, and it now also proposes to open them with vlf. Otherwise one can
open the file using `M-x vlf`.