Fixed the display of the project in modeline.
The bug was introduced with 31cc975 . One change was to load the theme
at the very beginning, because it's annoying to go back ot the default
UI when your config breaks. However now the theme is loaded before
Projectile. The fix is to move that code in init-powerline which is
loaded at the end.

Thank you git bissect.