Added flb-mode
M-x flb-mode toggles the frame-local-buffer mode, a minor mode that
makes any buffer selection function (such as C-x b) show only the
buffers that have been open in the current frame. In addition, when a
frame is closed the buffers that were only opened in that frame (and in
no other frames) are deleted automatically.

EXPERIMENTAL - use this at your own risk for now. It doesn't play well
with magit and it does not support bury-buffer. Maybe a better idea
would be to show all buffers but to reorder them according to frame
ownership?