Changed Helm key: C-c h => C-c M-h (or C-c H)
With this key, Helm first asks the user to select a Projectile project
and then searches for a file within the project. This is useful if you
want to open a file from a different project than the current one (the
one that is displayed in the modeline).

The previous keybinding was C-c h. It conflicts with Emacs, and is now
changed to C-c M-h (or C-c H).

A standard Emacs feature is that if you are looking for a particular key
and you know it starts with a given prefix, you can type the prefix
followed by C-h: Emacs will display the list of keys starting with that
prefix. So C-c C-h should display the list of keys starting with prefix
C-c.