Special feature for Julien to stay out of trouble
Add this in your after-init.el:

(setq markdown-command "/path/to/github-markup")

Then run this command while editing a markdown buffer:

M-x preview-github-markdown

This will open an eww buffer rendering the HTML output of the markdown
command.  The eww buffer does not need to be saved, you can just kill it
when you are done. It's exactly how GH will process it, but it's only as
good as eww can render it.

It's not real time but I'm working on it...