Use Helm as a default completion and selector
Add an option to use Helm as a replacement for IDO. By default it
is off. However when exordium-helm-everywhere is true Helm is used
by default.

Optionally allow for fuzzy matching when selection is narrowed
(default true).

Add a few keybindings to use Silver Searcher (Ag) with helm for
quick code scanning.