Added pref exordium-delete-trailing-whitespace
By default Exordium deletes all trailing whitespaces in a buffer before
saving it. Which is generally good practice. However some people work on
legacy projects that do not follow that rule, so this preference can be
used to disable this behavior.

Note: this should probably be turned into a minor mode, so that one can
turn off the behavior for some buffers only.