Add delete-trailing-whitespace-mode
Minor mode for deleting trailing whitespace upon saving a buffer. This
mode is enabled by default. If you want to disable this feature
globally, add this line in your ~/.emacs.d/prefs.el:

(setq 'exordium-delete-trailing-whitespace nil)

If you want to enable or disable the behavior for a particular buffer,
simply run M-x delete-trailing-white-space-mode. Note that the mode is
not displayed in the modeline, to avoid adding clutter into it.

If you want to enable or disable the behavior globally (for all
buffers), run M-x global-delete-trailing-whitespace-mode.