Add function exordium-toggle-rtags-cmake-mode
This can be used to turn off rtags-cmake-mode temporarily, when you are
doing a rebase for instance. The procedure is:

M-x rtags-stop
M-x rtags-toggle-rtags-cmake-mode
=> rtags-cmake-mode is DISABLED

After that, any new C++ file you open in a CMake project will not
trigger RTags to start, and powerline will show it.

To re-enable, simply do:

M-x rtags-toggle-rtags-cmake-mode
=> rtags-cmake-mode is ENABLED

After that, any new C++ file you open in a CMake project will restart
RTags, and powerline will show it.