Add company-mode as an alternative to auto-complete
This enables company-mode (globally) as an alternative to
auto-complete. While I haven't tested it extensively yet it
integrates with RTags pretty nicely and seems to work out of the box.

You can change which mode you want to use by setting
exordium-complete-mode to either :auto-complete or :company (or nil to
don't choose either).