Add option to copy both buffer regions in ediff
When solving merge conflicts with `ediff` I found myself more than once in a
need of having changes from both conflicting branches to be in final
merge. Sometimes they need refining, but usually parts of the conflicting
region were useful as is. IDK how to achieve that in a standard `ediff` without
selecting one change and then manually copying the other one.

This commit adds two keybindings to Ediff Control Panel: <kbd>A</kbd> and
<kbd>B</kbd>. They copy changes from both conflicting regions to the `C`
buffer. Two keys to allow for specifying order of yanks in a `C`.