Remove string manipulation functions
They are available in the minimum supported version of emacs-26.1.

Also rename the `string-truncate` function, so it is not colliding with
anything else, should it be added in a future versions. This also removes the
need for assertion of a function existence. At least this is what I guess the
assertion has been added back in 2016.