Support multiple frames and tabs for magit fullscreen
When entering a fullscreen mode for `magit` commands a current window
configuration is saved into a register. When the `magit` command is finished
the configuration is restored from the register.

Instead of using a single register a new name is generated for each tab (when
in `tab-bar-mode`) and each frame. That way, when there are multiple concurrent
`magit` sessions in an Emacs instance each of them is holding to its own window
configuration. With that, when a `magit` session is finished it restores a
window configuration that had been in a given frame/tab.