Configuration for emacs 29 treesitter support
Enable treesitter if it was compiled in, and provide the transition features
form treesit-auto, in particular auto install of -ts-mode if available and auto
fallback to the old mode. Arrange for this to be loaded late and copy the hooks
from commonly hooked modes into the -ts versions of the modes.