Use (if|when)-let\* functions
Seems like the star ending versions are preferred in community, while the
non-star ones may be deprecated (or not, discussion still pending), or even may
be subjected to change in semantic. For more explanation please read (rather
lengthy) discussion in https://debbugs.gnu.org/cgi/bugreport.cgi?bug=+73853.

My take from that is that the star version is safer bet.