Fix helm completion match selector
Current implementation missed setting `helm-completion-style` effectively using
`helm` as completion style, even though `exordium-helm-fuzzy-completion` has
been set.

I have been using `orderless` now, and meant to port it to Exordium, so used
the opportunity to do that.

The default is now `helm` (which was the default for a while now). Allowing to
select `helm-fuzzy` and `orderless`.