Remove RTags support
In the advent of Language Server Protocols (LSP) tools like `lsp-mode` and `eglot`
seem to be superior to what RTags has provided. Straight out of the box (and in
case of `eglot` that is a built-in it is literally straight out of the box)
they provide support to not only C++ but any language that has an LSP server
available.

Removing RTags support from Exordium doesn't mean it cannot be used with
Exordium. It simply means that we decided not to maintain extensions code any
more, as providing any reasonable support requires too much time.