Remove sqlite3 package
It's been required by package forge, only when there was no sqlite3 built-in
into Emacs. However, the former requires at Emacs-29 since Dec/2024 and
Emacs-29 has a possibility of having sqlite3 built-in. Let's assume sqlite3 is
built-in when needed (i.e., when installing forge for Emacs-29) and
don't bother with extra hacks.