feat(help)!: Display symbol details when completing helpful
When completing any of the `helpful` commands, for example
`helpful-function` (`C-h f`), `helpful-variable` or (`C-h v`) display details
of the completion. This is ahieved by using helm provided function, that is by
default used to provide affixations for symbols (i.e., when using
`describe-function` or `describe-variable`).

BREAKING CHANGE: Use built-in `completions-detailed` as per helm advice and
  hide it behind a `exodrium-help-extenstions` flag