fix(helm): Use original primary action when switching buffers
The default (first) action in `helm-source-buffers` is
`helm-buffer-switch-buffers`. With such a setup commands like
`switch-to-buffer-other-window` switch to buffer both in the current and in
the other window.

This commit fixes this by copying the original action (only the first one!)
from a source generated by `helm-comp-read`, followed by reminder of actions
specific to a buffer as defined in `helm-source-buffers`.

Also normalise sources before further processing. Although this is not strictly
necessary in current version of `helm`, this is more future proof.