refactor(treesit)!: Use treesit-auto recipes for forwarding hooks
The `treesit-auto` provides quite robust list of recipes for language grammars
and remappings that are used. Since the idea of a forwarding hooks is to
execute the non-ts-mode hooks in a ts-mode use that list to create hooks.

Also remove checks if both hooks exists. There's no issue when calling
`add-hook` with a non existing variable. Similarly  there is no issue when
calling `run-hooks` with a non existing one.

BREAKING CHANGE: Hooks are now added to all remappings defined in
 `treesit-auto`.