feat(helm): Add helm-wgrep
This brings a feature parity that has been lost when `helm-ag` become
obsolete. In short after a search, results can be exported to a grep buffer
(see actions in a helm buffer with search results by typing `TAB` or
`C-i`). Then such a buffer can be edited in `wgrep-mode` after typing `C-c
C-p`.

This should work not only in `helm-ag` search results, but also in
`helm-occur`, as well as in `grep` search results and perhaps other places.